requirements.txt
asgiref==3.2.7
django>=3.0.7
djangorestframework==3.11.0
django-redis==4.11.0
django-cors-headers==3.2.1
gunicorn==19.9.0
pytz==2019.3
sqlparse==0.3.1
requests==2.20.1
whitenoise==5.0.1
django-crispy-forms==1.9.1
